coverity#706094 Unintended sign extension
and

coverity#706093 Unintended sign extension
coverity#706092 Unintended sign extension
coverity#706091 Unintended sign extension

+++ svx/source/tbxctrls/colrctrl.cxx
@@ -365,7 +365,7 @@ void SvxColorDockingWindow::SetSize()
 
     // Scrollbar setzen/entfernen
     WinBits nBits = aColorSet.GetStyle();
-    if ( nLines * nCols >= nCount )
+    if ( static_cast<long>(nLines) * nCols >= nCount )
         nBits &= ~WB_VSCROLL;
     else
         nBits |= WB_VSCROLL;
@@ -528,7 +528,7 @@ void SvxColorDockingWindow::Resizing( Size& rNewSize )
 
     // Scrollbar setzen/entfernen
     WinBits nBits = aColorSet.GetStyle();
-    if ( nLines * nCols >= nCount )
+    if ( static_cast<long>(nLines) * nCols >= nCount )
         nBits &= ~WB_VSCROLL;
     else
         nBits |= WB_VSCROLL;
